requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
selenium==3.141.0
urllib3==1.25.3
Werkzeug==0.15.5
